% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARM2.R
\name{ARM2}
\alias{ARM2}
\title{Estimate NO2 with NOX/NO2 ratop}
\usage{
ARM2(nox, rmin = 0.5, rmax = 0.9)
}
\arguments{
\item{nox}{Concentration NOx in ug/m3}

\item{rmin}{Minimum ratio}

\item{rmax}{Maximum ratio}
}
\value{
A vector numeric with NO2 cconcnetration
}
\description{
Estimate NO2 with NOX/NO2 ratop
}
\examples{
no2 <- ARM2(1:10)
print(no2)

}
