% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_domian.R
\name{create_domian}
\alias{create_domian}
\title{Create domain and parameters to "gridcard" AERMOD}
\usage{
create_domian(coords, utm = 18, ext = NULL, res = NULL, pts = NULL)
}
\arguments{
\item{coords}{Numeric vector with coordinates center (X, Y)}

\item{utm}{Numeric value of UTM zone for DATUM WGS84}

\item{ext}{Numeric vector with extension in X and Y, or one numeric value to duplicated fo X and Y.}

\item{res}{Numeric vector with resolution in X and Y, or one numeric value to duplicated fo X and Y.}

\item{pts}{Numeric vector with points in X and Y, or one numeric value to duplicated fo X and Y.}
}
\value{
A list with parameters and shape object
}
\description{
Create a shape object (sfc) and parameters of GRIDCARD keyword to AERMOD.
}
\examples{
create_domian(coords=c(313860, 8617880), ext=4, res=50)

}
