% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_aersurface.R
\name{input_aersurface}
\alias{input_aersurface}
\title{Create input raster to AERSURFACE}
\usage{
input_aersurface(file, export)
}
\arguments{
\item{file}{Path file input}

\item{export}{Path file output}
}
\value{
A data frame
}
\description{
Import SURFACE output AERMET
}
\examples{
out_tif <- paste0(tempfile(), ".tif")
file_tif <- system.file("extdata", "land.tif", package="Raermod")
input_aersurface(file_tif, out_tif)
terra::rast(out_tif)

}
