% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapview_aermod_plt.R
\name{mapview_aermod_plt}
\alias{mapview_aermod_plt}
\title{Plot AERMOD.PLT output file}
\usage{
mapview_aermod_plt(
  plt,
  name_of_map_layer = "mapview_layer_name",
  string_filename = "export_filename",
  epsg = 32632,
  trans_factor = 1
)
}
\arguments{
\item{plt}{string ,path to the AERMOD plt file}

\item{name_of_map_layer}{string, name of the mapview layer}

\item{string_filename}{string,  name of the output files}

\item{epsg}{number, epsg to set crs in the raster object, default epsg 32632}

\item{trans_factor}{number, transformation factor to be applied to calculated values, leave it as default = 1}
}
\value{
mapview object exported as html and png files
}
\description{
Create mapview object from AERMOD.PLT file and export as html and png files
}
