% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_aermod_plt_recs.R
\name{stats_aermod_plt_recs}
\alias{stats_aermod_plt_recs}
\title{Statistics about discrete receptors in AERMOD.PLT output file}
\usage{
stats_aermod_plt_recs(plt)
}
\arguments{
\item{plt}{string, path to the AERMOD.PLT file}
}
\value{
tibble with: net_id, x and y coordinates, average concentration value, date expressed in the form 'yyyyjjjh'
}
\description{
Get relevant statistics for discrete receptors in AERMOD.PLT output file
}
