% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_aermod_plt.R
\name{read_aermod_plt}
\alias{read_aermod_plt}
\title{Read AERMOD.PLT output files as a tabular data}
\usage{
read_aermod_plt(plt, skip = 6)
}
\arguments{
\item{plt}{a string path to the AERMOD.PLT file}

\item{skip}{number of rows to skip, leave it as default = 6}
}
\value{
an object of type tibble
}
\description{
Returns a tibble of the output file AERMOD.PLT
}
